+++ tests/files/src/0896_probably_real.php
@@ -1,0 +1,10 @@
+<?php
+/** @param array $x */
+function test896($x) {
+    foreach ($x as $val) {
+        var_export($val);
+    }
+}
+test896(null);
+test896(new stdClass());
+test896(1);